# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/servicios": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all servicios",
                "description": "Display a listing of the clients.",
                "operationId": "23fd505e19d22862cb350bf22cc0894d",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filter by service name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paginacion",
                        "in": "query",
                        "description": "paginacion",
                        "required": false,
                        "schema": {
                            "type": "interger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/servicios/store": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "description": "Store a newly created client in storage.",
                "operationId": "e1743f49c3c85023982485419e069216",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "mantenimiento"
                                    },
                                    "precio": {
                                        "type": "string",
                                        "example": "100000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/servicios/{id}/show": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get a specific service by ID",
                "description": "Display the specified service.",
                "operationId": "040881aa4c31874acdad579e0e125a11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "servicio"
                                        },
                                        "precio": {
                                            "type": "string",
                                            "example": "10000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/api/servicios/{servicio}/update": {
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "update a new service",
                "description": "Update a newly created client in storage.",
                "operationId": "2a312d1cbe5cdbb6f788ba011bff4b86",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID of the service to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "mantenimiento"
                                    },
                                    "precio": {
                                        "type": "string",
                                        "example": "100000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/servicios/{servicio}": {
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service",
                "description": "Remove the specified service from storage.",
                "operationId": "e9e1a2118513dff16a3cc9617eacfd21",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID of the service to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}